Print tab character with more conventional method
This seems like a silly change to make, but it is in anticipation
of support for the `-v, --show-nonprinting` option. This option
will display nonprinting characters, relying on the Unicode code
point value to notate the character.

Since the tab (`^I`) character is nonprinting, it makes sense to
go ahead and refer to the Unicode code point value now in case
this bit of code is adopted into any new code written for the
nonprinting character code.