Properly parse long args, especially those with `=`
Long args were not being properly handled. In some cases long arguments
worked simply because both the short and long argument names began with
the same character, but even in that case, arguments in which the name
and value were separated with a `=` were not handled at all.

This fixes both the missing logic for parsing long arguments, and also
allows for separating argument names and values with `=` as suggested in
the help message.